requirements.txt
streamlit==1.28.0 ; python_version >= "3.10" and python_version < "3.11"
streamlit-player==0.1.5; python_version >= "3.10" and python_version < "3.11"
ultralytics==8.0.212; python_version >= "3.10" and python_version < "3.11"
opencv-python==4.8.1.78; python_version >= "3.10" and python_version < "3.11"
protobuf==3.20.*; python_version >= "3.10" and python_version < "3.11"
altair<5; python_version >= "3.10" and python_version < "3.11"
Pillow==9.5.0; python_version >= "3.10" and python_version < "3.11"
pytube==15.0.0; python_version >= "3.10" and python_version < "3.11"
lapx>=0.5.2; python_version >= "3.10" and python_version < "3.11"
Cython>=0.29.32; python_version >= "3.10" and python_version < "3.11"